+++ custom_components/fortios_kd/frontend/fortios-kd-dashboard-loader.js
@@ -2,3 +2,4 @@ import "./fortios-kd-dashboard.js";
 import "./fortios-kd-wifi-graphs-dashboard.js";
 import "./fortios-kd-arp-dashboard.js";
 import "./fortios-kd-arp-table-dashboard.js";
+import "./fortios-kd-dhcp-dashboard.js";